refactor: File and package structure of android-utils
