+++ features/resume/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/resume/internal/ProveYourIdentityEntryPointsImplTest.kt
@@ -12,7 +12,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import uk.gov.android.ui.theme.m3.GdsTheme
 import uk.gov.logging.testdouble.analytics.FakeAnalyticsLogger
-import uk.gov.onelogin.criorchestrator.libraries.androidutils.FakeResourceProvider
+import uk.gov.onelogin.criorchestrator.libraries.androidutils.resources.FakeResourceProvider
 
 @RunWith(AndroidJUnit4::class)
 class ProveYourIdentityEntryPointsImplTest {
