+++ features/resume/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/resume/internal/ProveYourIdentityViewModelTest.kt
@@ -10,7 +10,7 @@ import uk.gov.logging.api.v3dot1.logger.logEventV3Dot1
 import uk.gov.logging.api.v3dot1.model.AnalyticsEvent
 import uk.gov.logging.api.v3dot1.model.RequiredParameters
 import uk.gov.logging.api.v3dot1.model.TrackEvent
-import uk.gov.onelogin.criorchestrator.libraries.androidutils.FakeResourceProvider
+import uk.gov.onelogin.criorchestrator.libraries.androidutils.resources.FakeResourceProvider
 
 class ProveYourIdentityViewModelTest {
     private val analyticsLogger = mock<AnalyticsLogger>()
