+++ libraries/android-utils/build.gradle.kts
@@ -1,9 +1,17 @@
+import com.android.build.gradle.LibraryExtension
+
 plugins {
     listOf(
         "uk.gov.onelogin.criorchestrator.android-lib-config",
         "uk.gov.onelogin.criorchestrator.local-ui-test-config",
     ).forEach {
         id(it)
+    }
+}
+
+configure<LibraryExtension> {
+    testFixtures {
+        androidResources = true
     }
 }
