+++ libraries/android-utils/src/main/kotlin/uk/gov/onelogin/criorchestrator/libraries/androidutils/resources/AndroidResourceProvider.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.libraries.androidutils
+package uk.gov.onelogin.criorchestrator.libraries.androidutils.resources
 
 import android.content.Context
 import androidx.annotation.StringRes
@@ -6,19 +6,6 @@ import com.squareup.anvil.annotations.ContributesBinding
 import uk.gov.logging.api.analytics.extensions.getEnglishString
 import uk.gov.onelogin.criorchestrator.libraries.di.CriOrchestratorScope
 import javax.inject.Inject
-
-/**
- * Interface for providing resources.
- */
-fun interface ResourceProvider {
-    /**
-     * Function that returns the English versions of strings for non-user-facing usage,
-     * typically for analytics.
-     */
-    fun getEnglishString(
-        @StringRes resId: Int,
-    ): String
-}
 
 @ContributesBinding(CriOrchestratorScope::class)
 class AndroidResourceProvider
