+++ libraries/android-utils/src/test/kotlin/uk/gov/onelogin/criorchestrator/libraries/androidutils/resources/AndroidResourceProviderTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.libraries.androidutils
+package uk.gov.onelogin.criorchestrator.libraries.androidutils.resources
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
@@ -6,7 +6,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-import uk.gov.onelogin.criorchestrator.libraries.androidutils.test.R as testR
+import uk.gov.onelogin.criorchestrator.libraries.androidutils.testFixtures.R as testR
 
 @RunWith(AndroidJUnit4::class)
 class AndroidResourceProviderTest {
