+++ libraries/android-utils/src/testFixtures/kotlin/uk/gov/onelogin/criorchestrator/libraries/androidutils/resources/FakeResourceProvider.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.libraries.androidutils
+package uk.gov.onelogin.criorchestrator.libraries.androidutils.resources
 
 class FakeResourceProvider : ResourceProvider {
     override fun getEnglishString(resId: Int): String = "dummy string"
