+++ libraries/android-utils/src/testFixtures/res/values/strings.xml
